automatically write an html report summary at end of run
