+++ .gitignore
@@ -164,3 +164,4 @@ garak.*.jsonl
 garak.log
 hitlog.*.jsonl
 .vscode
+logs/
