+++ garak/analyze/count_tokens.py
