+++ garak/analyze/templates/digest_detector.jinja
